add copy private key component
